requirements.txt
pillow >= 8.2.0
streamlit >= 0.82.0
tensorflow  >=2.5.0
numpy >= 1.19.5